Move exception handler to starter module.
