+++ agrest-spring-boot-starter/src/main/java/io/agrest/spring/starter/AgrestProperties.java
@@ -1,0 +1,18 @@
+package io.agrest.spring.starter;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "agrest")
+public class AgrestProperties {
+
+  private boolean enabled = false;
+
+
+  public boolean isEnabled() {
+    return enabled;
+  }
+
+  public void setEnabled(boolean enabled) {
+    this.enabled = enabled;
+  }
+}
