+++ agrest-spring-boot-starter/src/main/java/io/agrest/spring/starter/exceptions/RestResponseEntityExceptionHandler.java
@@ -1,7 +1,8 @@
-package io.agrest.spring.it;
+package io.agrest.spring.starter.exceptions;
 
 import io.agrest.AgException;
 
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -9,6 +10,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import static org.springframework.core.Ordered.LOWEST_PRECEDENCE;
 
 @ControllerAdvice
 public class RestResponseEntityExceptionHandler extends ResponseEntityExceptionHandler {
