+++ agrest-spring-it/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-agrest.enabled = true
+agrest.enabled=true
